fix: Page Count Calculation
For both team and user donations, the page count was calculated by dividing total donations by 100 instead of the limit passed by the user.  While in the URL generation, offset is based on page multiplied by passed limit.  Because of the difference, the `countPages` would be incorrect for users who use non-100 limits.

This change corrects that and sets page calculation based on the user-defined limit.